# Improvements to KernelAbstractions code?