--- a/src/macro.jl
+++ b/src/macro.jl
@@ -974,14 +974,15 @@ function make_many_actors(act!, args, ex1, outer::Vector, ex3, inner::Vector, ex
                 outer = [Symbol(EPS, 1)] # fake index name, only appears in @index
                 sizes = [:(one(Int))]    # iterate over 1:1
             end
-            const_args = map(args) do a
-                a isa Symbol || return a  # this skips output ZED::AbstractArray{TYP}
-                a == store.leftarray && return a  # case A[i] = A[i]^2 / B[i,j]
-                :(@Const($a))
-            end
+            # const_args = map(args) do a
+            #     a isa Symbol || return a  # this skips output ZED::AbstractArray{TYP}
+            #     a == store.leftarray && return a  # case A[i] = A[i]^2 / B[i,j]
+            #     :(@Const($a))
+            # end
             kex1 = quote
-
-                KernelAbstractions.@kernel function $kernel($(const_args...), @Const($KEEP), @Const($FINAL)) where {$TYP}
+                # @Const removed, see https://github.com/mcabbott/Tullio.jl/pull/32
+                # KernelAbstractions.@kernel function $kernel($(const_args...), @Const($KEEP), @Const($FINAL)) where {$TYP}
+                KernelAbstractions.@kernel function $kernel($(args...), $KEEP, $FINAL) where {$TYP}
                     ($(outer...),) = @index(Global, NTuple)
                     ($ex1; $ex3; $ex4; $ex6)
                 end