@merckxiaan 2020-09-07
This is what I get with the `try 0` commit
```
julia> [f(a) for _ in 1:20]
ERROR: Waiting on the GPU for an CPU event to finish is currently not supported.
We have encountered deadlocks arising, due to interactions with the CUDA
driver. If you are certain that you are deadlock free, you can use `unsafe_wait`
instead.

Stacktrace:
 [1] error(::String) at .\error.jl:33
 [2] wait(::CUDADevice, ::KernelAbstractions.CPUEvent, ::Function, ::CuStream) at C:\Users\jules\.julia\packages\KernelAbstractions\bIZIO\src\backends\cuda.jl:91
 [3] wait(::CUDADevice, ::MultiEvent{Tuple{KernelAbstractions.CPUEvent}}, ::Function, ::CuStream) at C:\Users\jules\.julia\packages\KernelAbstractions\bIZIO\src\backends\cuda.jl:83
 [4] (::KernelAbstractions.Kernel{CUDADevice,KernelAbstractions.NDIteration.StaticSize{(256,)},KernelAbstractions.NDIteration.DynamicSize,var"#gpu_##��#352#3"})(::CuArray{Float32,2}, ::Vararg{Any,N} where N; ndrange::Tuple{Int64,Int64}, dependencies::KernelAbstractions.CPUEvent, workgroupsize::Nothing, progress::Function) at C:\Users\jules\.julia\packages\KernelAbstractions\bIZIO\src\backends\cuda.jl:206
 [5] ���! at C:\Users\jules\.julia\packages\Tullio\uvqiL\src\macro.jl:1000 [inlined]
 [6] ���! at C:\Users\jules\.julia\packages\Tullio\uvqiL\src\macro.jl:997 [inlined]
 [7] threader(::var"#���!#1", ::Type{CuArray{Float32,2}}, ::CuArray{Float32,2}, ::Tuple{CuArray{Float32,2},CuArray{Float32,2}}, ::Tuple{Base.OneTo{Int64},Base.OneTo{Int64}}, ::Tuple{Base.OneTo{Int64}}, ::Function, ::Int64, ::Nothing) at C:\Users\jules\.julia\packages\Tullio\uvqiL\src\eval.jl:89
 [8] (::var"#ℳ��ℯ#10"{var"#���!#1"})(::CuArray{Float32,2}, ::CuArray{Float32,2}) at C:\Users\jules\.julia\packages\Tullio\uvqiL\src\macro.jl:802
 [9] (::Tullio.Eval{var"#ℳ��ℯ#10"{var"#���!#1"},var"#265#∇ℳ��ℯ#9"{var"#∇���!#5"}})(::CuArray{Float32,2}, ::Vararg{CuArray{Float32,2},N} where N) at C:\Users\jules\.julia\packages\Tullio\uvqiL\src\eval.jl:20
 [10] f(::CuArray{Float32,2}) at .\REPL[3]:1
 [11] (::var"#12#13")(::Int64) at .\none:0
 [12] iterate at .\generator.jl:47 [inlined]
 [13] collect(::Base.Generator{UnitRange{Int64},var"#12#13"}) at .\array.jl:686
 [14] top-level scope at REPL[4]:1
```