@merckxiaan 2020-09-07
I also ran right before `try 0` and commented here https://github.com/JuliaGPU/KernelAbstractions.jl/issues/108.
Trying out `CuDevice suggestions` gives the same error as `add Const to kernel arguments`.
I've got no clue about what's going on and I'm also not sure I can easily combine branches to test?