@mcabbott 2020-09-07
If feeling energetic, you could `] dev` this package and mess around locally?

`CuDevice` doesn't seem to show up in https://github.com/JuliaGPU/KernelAbstractions.jl/tree/master/examples, maybe it's new, I'm not sure. Maybe https://github.com/JuliaGPU/CUDA.jl/search?q=CuDevice&unscoped_q=CuDevice has clues.