@merckxiaan 2020-09-07
With my luck, the monkeys would've finished Shakespeare's work already.
I replaced `CuDevice(0)` with `CUDADevice()` and everything magically works!

As for making stuff const, I quickly ran Mason's CUDA example mentioned in #30 but there's seems to be no aprreciable increase in performance.
before `const`-fix (yet after `CUDADevice`-fix):
```
N = 2
  0.000379 seconds (133 CPU allocations: 4.453 KiB) #generated by Tullio
  0.000177 seconds (15 CPU allocations: 416 bytes) #CUBLAS
N = 10
  0.000506 seconds (167 CPU allocations: 4.984 KiB)
  0.000056 seconds (15 CPU allocations: 416 bytes) 
N = 50
  0.000216 seconds (133 CPU allocations: 4.453 KiB)
  0.000102 seconds (15 CPU allocations: 416 bytes)
N = 100
  0.000198 seconds (133 CPU allocations: 4.453 KiB)
  0.000061 seconds (15 CPU allocations: 416 bytes)
N = 200
  0.000521 seconds (133 CPU allocations: 4.453 KiB)
  0.000390 seconds (15 CPU allocations: 416 bytes)
N = 500
  0.002040 seconds (133 CPU allocations: 4.453 KiB)
  0.000284 seconds (15 CPU allocations: 416 bytes)
N = 1000
  0.014123 seconds (133 CPU allocations: 4.453 KiB)
  0.000747 seconds (15 CPU allocations: 416 bytes)
N = 2000
  0.119267 seconds (133 CPU allocations: 4.453 KiB)
  0.004502 seconds (15 CPU allocations: 416 bytes)
N = 5000
  2.391341 seconds (133 CPU allocations: 4.453 KiB)
  0.045763 seconds (15 CPU allocations: 416 bytes)
N = 10000
 19.480897 seconds (133 CPU allocations: 4.453 KiB)
  0.437446 seconds (15 CPU allocations: 416 bytes)
```

after `const`-fix:
```
N = 2
  0.000229 seconds (133 CPU allocations: 4.453 KiB)
  0.000291 seconds (15 CPU allocations: 416 bytes)
N = 10
  0.000325 seconds (133 CPU allocations: 4.453 KiB)
  0.000267 seconds (15 CPU allocations: 416 bytes)
N = 50
  0.000933 seconds (133 CPU allocations: 4.453 KiB)
  0.000186 seconds (15 CPU allocations: 416 bytes)
N = 100
  0.000517 seconds (133 CPU allocations: 4.453 KiB)
  0.000209 seconds (15 CPU allocations: 416 bytes)
N = 200
  0.000422 seconds (133 CPU allocations: 4.453 KiB)
  0.000278 seconds (15 CPU allocations: 416 bytes)
N = 500
  0.001739 seconds (133 CPU allocations: 4.453 KiB)
  0.000212 seconds (15 CPU allocations: 416 bytes)
N = 1000
  0.014014 seconds (133 CPU allocations: 4.453 KiB)
  0.000805 seconds (15 CPU allocations: 416 bytes)
N = 2000
  0.116392 seconds (133 CPU allocations: 4.453 KiB)
  0.004565 seconds (15 CPU allocations: 416 bytes)
N = 5000
  2.377929 seconds (133 CPU allocations: 4.453 KiB)
  0.044224 seconds (15 CPU allocations: 416 bytes)
N = 10000
 19.390380 seconds (133 CPU allocations: 4.453 KiB)
  0.434024 seconds (15 CPU allocations: 416 bytes)
```