@mcabbott 2020-09-08
Great, well spotted. 

Does this version fix https://github.com/JuliaGPU/KernelAbstractions.jl/issues/108 too? 

Thanks for checking the benchmark. I guess more complicated tricks are needed there somehow.