@merckxiaan 2020-09-08
> Does this version fix [JuliaGPU/KernelAbstractions.jl#108](https://github.com/JuliaGPU/KernelAbstractions.jl/issues/108) too?

It does!

 > CI is still unhappy it seems: https://gitlab.com/JuliaGPU/Tullio.jl/-/jobs/725753381

When I run these tests locally without the `const`-changes (but with `CUDADevice`), they pass.